requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
pysnmp==4.4.12
paramiko==3.4.0
pandas==2.2.0
openpyxl==3.1.2
tqdm==4.66.2
mysql-connector-python==9.0.0  # For MySQL support
